Fixes input type checks (#8147)

+++ integration-tests/testsetups/keeper_benchmark.go
@@ -349,12 +349,12 @@ func (k *KeeperBenchmarkTest) ensureInputValues(t *testing.T) {
 	require.NotNil(t, inputs.KeeperRegistrySettings, "You need to set KeeperRegistrySettings")
 	require.NotNil(t, k.Inputs.ChainlinkNodeFunding, "You need to set a funding amount for chainlink nodes")
 	clFunds, _ := k.Inputs.ChainlinkNodeFunding.Float64()
-	require.GreaterOrEqual(t, clFunds, 0, "Expecting Chainlink node funding to be more than 0 ETH")
-	require.Greater(t, inputs.CheckGasToBurn, 0, "You need to set an expected amount of gas to burn on checkUpkeep()")
+	require.GreaterOrEqual(t, clFunds, 0.0, "Expecting Chainlink node funding to be more than 0 ETH")
+	require.Greater(t, inputs.CheckGasToBurn, int64(0), "You need to set an expected amount of gas to burn on checkUpkeep()")
 	require.GreaterOrEqual(
 		t, inputs.KeeperRegistrySettings.CheckGasLimit, inputs.CheckGasToBurn, "CheckGasLimit should be >= CheckGasToBurn",
 	)
-	require.Greater(t, inputs.PerformGasToBurn, 0, "You need to set an expected amount of gas to burn on performUpkeep()")
+	require.Greater(t, inputs.PerformGasToBurn, int64(0), "You need to set an expected amount of gas to burn on performUpkeep()")
 	require.NotNil(t, inputs.UpkeepSLA, "Expected UpkeepSLA to be set")
 	require.NotNil(t, inputs.FirstEligibleBuffer, "You need to set FirstEligibleBuffer")
 	require.NotNil(t, inputs.RegistryVersions[0], "You need to set RegistryVersion")
